requirements/base.txt
django==2.0.5
pytz==2018.4
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
Pillow==5.1.0
